Tile-server/requirements.txt
affine==2.4.0
annotated-types==0.7.0
anyio==3.7.1
attrs==25.1.0
certifi==2025.1.31
click==8.1.8
click-plugins==1.1.1
cligj==0.7.2
contourpy==1.3.0
cycler==0.12.1
exceptiongroup==1.2.2
fastapi==0.104.1
fonttools==4.56.0
h11==0.14.0
idna==3.10
importlib_resources==6.5.2
kiwisolver==1.4.7
matplotlib==3.9.4
mercantile==1.2.1
numpy==1.24.4
packaging==24.2
Pillow==10.0.1
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-multipart==0.0.6
rasterio==1.3.8
six==1.17.0
sniffio==1.3.1
snuggs==1.4.7
starlette==0.27.0
typing_extensions==4.12.2
uvicorn==0.23.2
zipp==3.21.0
